as.big.data.table {big.data.table}	R Documentation

Convert to big.data.table

Description

Convert to big.data.table

Usage

as.big.data.table(x, ...)

## S3 method for class 'function'
as.big.data.table(x, rscl, partition.by, partitions,
  parallel = TRUE, ..., .log = getOption("bigdatatable.log", FALSE))

## S3 method for class 'call'
as.big.data.table(x, rscl, partition.by, partitions,
  parallel = TRUE, ..., .log = getOption("bigdatatable.log", FALSE))

## S3 method for class 'list'
as.big.data.table(x, partition.by, partitions, parallel = TRUE,
  ..., .log = getOption("bigdatatable.log", FALSE))

## S3 method for class 'data.table'
as.big.data.table(x, rscl, partition.by, partitions,
  parallel = TRUE, ..., .log = getOption("bigdatatable.log", FALSE))

Arguments

x
object to cast into big.data.table, can be data.table, function, quoted call or a list.

rscl
list of Rserve connections, can be validated by 'is.rscl(rscl, silent=FALSE)'.

partition.by
character vector of column names to be used for partitioning, 'uniqueN' by those columns should be lower than number of nodes.

partitions
data.table of unique combinations of values in *partition.by* columns.

parallel
logical, see '?bdt.eval'.

.log
logical if *TRUE* then logging will be done using logR to postgres db.

...
arguments passed to methods.

Value

big.data.table object.

Note

Supported 'x' data types are *data.table* (will be automatically spread across the nodes), *function* or quoted *call* will be evaluated on each node and assigned to 'x' variable, a *list* will struct big.data.table on already working set of nodes, all having 'x' data.tables.

[Package big.data.table version 0.3.3 Index]