bdt.assign {big.data.table}	R Documentation

big.data.table assign object

Description

Saves the object to nodes, handles chunking/partitioning.

Usage

bdt.assign(x, name, value, parallel = TRUE,
  .log = getOption("bigdatatable.log", FALSE))

Arguments

x
big.data.table.

name
character variable name to which assign *value* in each node, for *x* data.table it should be equal to 'x'.

value
an R object to save on node, if it is data.table then it will be partitioned into chunks, if not partition defined it will make equal rows chunks.

parallel
logical if parallel *TRUE* (default) it will send expression to nodes using 'wait=FALSE' and collect results afterward executing each node in parallel.

.log
currently ignored.

[Package big.data.table version 0.3.3 Index]