bdt.eval {big.data.table}	R Documentation

big.data.table evaluate on nodes

Description

Main engine for passing queries to nodes, control parallelization, rbinding or simplifing returned object. Allows to measure timing and verbose messages.

Usage

bdt.eval(x, expr, lazy = TRUE, send = FALSE, simplify = TRUE,
  rbind = TRUE, parallel = TRUE, silent = TRUE,
  .log = getOption("bigdatatable.log", FALSE))

Arguments

x
big.data.table.

expr
expression.

lazy
logical if TRUE then *expr* is substituted.

send
logical, if TRUE submit expression appended with 'TRUE' to not fetch potentially big results from provided *expr*, useful for data.table *set** or ':=' functions.

simplify
logical if *TRUE* (default) it will simplify list of 1 length same type objects to vector.

rbind
logical if *TRUE* (default) results are data.table they will be rbinded.

parallel
logical if parallel *TRUE* (default) it will send expression to nodes using 'wait=FALSE' and collect results afterward executing each node in parallel.

silent
should be silently catched by 'try' or 'logR' if enabled.

.log
logical if *TRUE* then logging will be done using logR to postgres db.

Value

Depending on *simplify, rbind* the results of evaluated expression.

[Package big.data.table version 0.3.3 Index]