rscl.eval {big.data.table}	R Documentation

Eval on list of Rserve instances

Description

A wrapper to 'RS.eval'.

Usage

rscl.eval(rscl = getOption("bigdatatable.rscl"), x, wait = TRUE,
  lazy = TRUE, parallel = FALSE, simplify = TRUE)

Arguments

rscl
lists of Rserve connections.

x
expression, can be quoted then use *lazy* arg. Can be list of quoted expression to run on all nodes in by expression batches on all R nodes.

wait
logical default TRUE passed to 'RS.eval'. If *parallel* arg is TRUE then this argument is ignored.

lazy
logical default TRUE, when FALSE then *x* is quoted or a list of quoted.

parallel
logical, default FALSE, when TRUE it will auto collect results

simplify
logical, default TRUE, passed to underlying 'sapply'.

Value

Logical matrix.

[Package big.data.table version 0.3.3 Index]