[.big.data.table {big.data.table}	R Documentation

Subset from big.data.table

Description

Subset from big.data.table

Usage

## S3 method for class 'big.data.table'
x[..., new.var, new.copy = FALSE,
  parallel = TRUE,
  outer.aggregate = getOption("bigdatatable.outer.aggregate", FALSE),
  .log = getOption("bigdatatable.log", FALSE)]

Arguments

x
big.data.table object.

new.var
character scalar, name of new variable where query results should be cached.

new.copy
logical if *TRUE* it will make deep copy while saving to *new.var*.

parallel
logical if parallel *TRUE* (default) it will send expression to nodes using 'wait=FALSE' and collect results afterward executing each node in parallel.

outer.aggregate
logical or a function, if *TRUE* will able the same query to rbind of results from each node, should not be used with '.SD', '.N', etc. Also conflicts with filtering in 'i'. Can be also a function taking first argument 'x' rbinded data.table.

.log
logical if *TRUE* then logging will be done using logR to postgres db.

...
arguments passed to each node '[.data.table' call: *i, j, by, keyby...*.

Value

data.table object.

Note

Results from nodes are rbinded and the same call is evalated on combined results. That means the column names cannot be renamed or simplified to vector in '...' call. Use '[[.big.data.table' for deeper flexibility.

[Package big.data.table version 0.3.3 Index]