[[.big.data.table {big.data.table}	R Documentation

Extract from big.data.table

Description

Extract from big.data.table

Usage

## S3 method for class 'big.data.table'
x[[j, expr, lazy = TRUE, send = FALSE, i, ...,
  simplify = TRUE, rbind = TRUE, parallel = TRUE,
  .log = getOption("bigdatatable.log", FALSE)]]

Arguments

x
big.data.table object.

j
numeric scalar, if provided then all other arguments are ignored and subset behaves the same way as for data.table, but returns 0 length column.

expr
expression to be evaluated on node where data.table objects are stored as 'x' variable in '.GlobalEnv'.

lazy
logical if TRUE then *expr* is substituted.

send
logical, if TRUE submit expression appended with 'TRUE' to not fetch potentially big results from provided *expr*, useful for data.table __set*__ or ':=' functions.

i
numeric restrict expression to particular nodes

simplify
logical passed to 'bdt.eval', affects the type of returned object.

rbind
logical passed to 'bdt.eval', affects the type of returned object.

parallel
logical if parallel *TRUE* (default) it will send expression to nodes using 'wait=FALSE' and collect results afterward executing each node in parallel.

.log
logical if *TRUE* then logging will be done using logR to postgres db.

...
ignored.

Value

When using *j* arg the 0 length variable from underlying data is returned. Otherwise the results from expression evaluated as *lapply*. When using *rbind* or *simplify* the returned list be can simplified.

[Package big.data.table version 0.3.3 Index]